feat: 카메라로 찍힌 음식 다른 뷰로 넘겨주기 기능 구현
- Models -> ChosenFood 모델에 선택된 음식 String 형태로 저장
- 폴더 구조 변경(Views, Models, Util->ViewAssets 추가)
- FeedButton -> FoodResultView로 넘기기 성공
- MainView 화면의 구린 디노 예쁜 디노로 교체
- MainView 화면의 도감, 먹이주기 버튼 이미지로 교체
