+++ DinoDina/DinoDina/Assets.xcassets/Dina/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
